LPD-57955 Merge SearchLayoutFactory/SearchLayoutFactoryImpl into a pojo
