+++ modules/apps/portal-search/portal-search-web/src/main/java/com/liferay/portal/search/web/internal/layout/prototype/AddLayoutPrototypeInitialRequestPortalInstanceLifecycleListener.java
@@ -12,6 +12,10 @@ import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.GroupConstants;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.service.GroupLocalService;
+import com.liferay.portal.kernel.service.LayoutLocalService;
+import com.liferay.portal.kernel.service.LayoutPrototypeLocalService;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.util.Localization;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.service.component.annotations.Activate;
@@ -30,24 +34,28 @@ public class AddLayoutPrototypeInitialRequestPortalInstanceLifecycleListener
 	@Override
 	protected void activate(BundleContext bundleContext) {
 		super.activate(bundleContext);
+
+		_searchLayoutFactory = new SearchLayoutFactory(
+			_groupLocalService, _layoutLocalService,
+			_layoutPrototypeLocalService, _localization, _userLocalService);
 	}
 
 	@Override
 	protected void doPortalInstanceRegistered(long companyId) throws Exception {
-		Layout layout = searchLayoutFactory.createSearchLayoutPrototype(
+		Layout layout = _searchLayoutFactory.createSearchLayoutPrototype(
 			companyId);
 
 		if (layout == null) {
 			return;
 		}
 
-		Group guestGroup = groupLocalService.getGroup(
+		Group guestGroup = _groupLocalService.getGroup(
 			companyId, GroupConstants.GUEST);
 
 		try {
 			MergeLayoutPrototypesThreadLocal.setInProgress(true);
 
-			searchLayoutFactory.createSearchLayout(guestGroup);
+			_searchLayoutFactory.createSearchLayout(guestGroup);
 		}
 		finally {
 			MergeLayoutPrototypesThreadLocal.setInProgress(false);
@@ -55,9 +63,20 @@ public class AddLayoutPrototypeInitialRequestPortalInstanceLifecycleListener
 	}
 
 	@Reference
-	protected GroupLocalService groupLocalService;
+	private GroupLocalService _groupLocalService;
 
 	@Reference
-	protected SearchLayoutFactory searchLayoutFactory;
+	private LayoutLocalService _layoutLocalService;
+
+	@Reference
+	private LayoutPrototypeLocalService _layoutPrototypeLocalService;
+
+	@Reference
+	private Localization _localization;
+
+	private SearchLayoutFactory _searchLayoutFactory;
+
+	@Reference
+	private UserLocalService _userLocalService;
 
 }
